backend/requirements.txt
fastapi
uvicorn
requests
faiss-cpu
numpy
python-dotenv
